Add forge_gui, self-play components, MoE routers, and key reorganization
- Add forge_gui/ package + launch_gui.py entry point for desktop UI
- Add self-play subsystem: data_quality, grpo_trainer, io_match,
  live_status, monitoring, replay_buffer
- Add MoE keyword_router and semantic_router; expand airmoe_infinite
- Add spectral_injection_key; reorganize keys/ into 12 category subdirs
  with per-subdir __init__.py re-exports
- Add HF dataset download/inject scripts (coding, math, reasoning, etc.)
- Add unit tests for moe, keyword_router, self_play components
- Remove obsolete web/ assets, debug scripts, and v4 baking pipeline
- Refresh AGENTS.md and scratchpad with current architecture notes

+++ research/moe/moe.py
@@ -87,7 +87,7 @@ class Router(nn.Module):
             expert_idx = top_k_indices[:, k]  # (N,)
             weight = top_k_weights[:, k]  # (N,)
             dispatch_mask.scatter_(1, expert_idx.unsqueeze(1), 1.0)
-            gating_weights.scatter_(1, expert_idx.unsqueeze(1), weight.unsqueeze(1))
+            gating_weights.scatter_(1, expert_idx.unsqueeze(1), weight.unsqueeze(1).to(gating_weights.dtype))
 
         # Load balancing loss (encourages uniform expert utilization).
         # Following Switch Transformer: loss = n_experts * sum(mean_gate * mean_dispatch)
